Move some lockfile helpers to `core/goals/generate_lockfiles.py`
No changes made. Just moved around.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
