+++ src/python/pants/backend/python/subsystems/python_tool_base.py
@@ -16,13 +16,11 @@ from pants.backend.python.util_rules.pex import (
     ToolCustomLockfile,
     ToolDefaultLockfile,
 )
+from pants.core.goals.generate_lockfiles import DEFAULT_TOOL_LOCKFILE, NO_TOOL_LOCKFILE
 from pants.engine.fs import FileContent
 from pants.option.errors import OptionsError
 from pants.option.subsystem import Subsystem
 from pants.util.ordered_set import FrozenOrderedSet
-
-DEFAULT_TOOL_LOCKFILE = "<default>"
-NO_TOOL_LOCKFILE = "<none>"
 
 
 class PythonToolRequirementsBase(Subsystem):
